requirements.txt
# scan2sim/requirements.txt
# python 3.11 / pip 25.3 / torch 2.9.0+cu130

# python -m pip install --no-build-isolation -r requirements.txt

# --- core ---
numpy>=2.0,<2.1
scipy>=1.15,<2
pandas>=2.2.2,<3
scikit-learn>=1.7,<2
tqdm>=4.67.1,<5
pyyaml>=6.0.3,<7
yacs>=0.1.8,<0.2
rich>=14.2.0,<15

# --- plotting / analysis ---
matplotlib>=3.9.0,<4
seaborn>=0.13.2,<0.14

# --- point cloud / lidar IO ---
open3d>=0.19.0,<0.20
laspy>=2.7,<3
pyarrow>=22,<24

# --- ouster lidar sdk ---
ouster-sdk==0.16.0

# --- body model ---
smplx==0.1.28

# --- legacy (chumpy) ---
# chumpy: Python 3.11 getargspec fix commit
chumpy @ git+https://github.com/mattloper/chumpy.git@9b045ff5d6588a24a0bab52c83f032e2ba433e17
six>=1.16,<2

# --- dev ---
pytest>=9,<10
pytest-cov>=7,<8
ruff>=0.14,<0.15
pre-commit>=4.5,<5
ipykernel>=7.1,<8
jupyterlab>=4.4,<4.5
